Re-organize structure
