+++ chainer/iterators/__init__.py
@@ -3,5 +3,5 @@ from chainer.iterators.multiprocess_iterator import MultiprocessIterator  # NOQA
 from chainer.iterators.multithread_iterator import MultithreadIterator  # NOQA
 from chainer.iterators.serial_iterator import SerialIterator  # NOQA
 
-from chainer.iterators.order_samplers import OrderSampler  # NOQA
-from chainer.iterators.order_samplers import ShuffleOrderSampler  # NOQA
+from chainer.iterators.order_samplers.order_sampler import OrderSampler  # NOQA
+from chainer.iterators.order_samplers.shuffle_order_sampler import ShuffleOrderSampler  # NOQA
