+++ chainer/iterators/order_samplers/__init__.py
@@ -1,0 +1,2 @@
+from chainer.iterators.order_samplers.order_sampler import OrderSampler  # NOQA
+from chainer.iterators.order_samplers.shuffle_order_sampler import ShuffleOrderSampler  # NOQA
